Restructured using: 	router for controller endpoints 	schemas to define objects
